Removed scheduled jobs and added RabbitService
