+++ src/Application/Interfaces/Services/IRabbitService.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace VoteApp.Application.Interfaces.Services
+{
+    internal interface IRabbitService
+    {
+        Task SendAsync();
+    }
+}
