--- src/Infrastructure/Services/Rabbit/RabbitMqConfiguration.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace VoteApp.Application.Interfaces.Rabbit
-{
-    public class RabbitMqConfiguration
-    {
-
-        public string HostName { get; set; }
-        public string QueueName { get; set; }
-        public string UserName { get; set; }
-        public string Password { get; set; }
-    }
-}
