+++ src/Server/Startup.cs
@@ -14,8 +14,6 @@ using System.IO;
 using VoteApp.Server.Filters;
 using VoteApp.Server.Managers.Preferences;
 using Microsoft.Extensions.Localization;
-using VoteApp.Application.Interfaces.Rabbit;
-using VoteApp.Application.RabbitJobScheduler;
 
 namespace VoteApp.Server
 {
@@ -68,17 +66,6 @@ namespace VoteApp.Server
             });
             services.AddLazyCache();
 
-
-
-            //Test av rabbitmq/hangfire
-            services.AddScoped<IRabbitJob, RabbitJob>();
-
-            
-            services.AddOptions();
-            services.Configure<RabbitMqConfiguration>(_configuration.GetSection("RabbitMq"));
-            services.AddTransient<IRabbitJob, RabbitJob>();
-            
-
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IStringLocalizer<Startup> localizer)
@@ -112,8 +99,7 @@ namespace VoteApp.Server
                 Authorization = new[] { new HangfireAuthorizationFilter() }
             });
 
-            HangfireJobScheduler.ScheduleReccuringJobs();
-
+           
             app.UseEndpoints();
             app.ConfigureSwagger();
             app.Initialize(_configuration);
